+++ tests/dummy/app/JsonApi/Videos/Schema.php
@@ -18,7 +18,6 @@
 namespace DummyApp\JsonApi\Videos;
 
 use CloudCreativity\LaravelJsonApi\Schema\EloquentSchema;
-use DummyApp\Video;
 
 class Schema extends EloquentSchema
 {
@@ -34,19 +33,11 @@ class Schema extends EloquentSchema
     protected $attributes = null;
 
     /**
-     * @param Video $resource
-     * @param bool $isPrimary
-     * @param array $includeRelationships
-     * @return array
+     * @var array
      */
-    public function getRelationships($resource, $isPrimary, array $includeRelationships)
-    {
-        return [
-            'uploaded-by' => [
-                self::DATA => isset($includeRelationships['uploaded-by']) ?
-                    $resource->user : $this->createBelongsToIdentity($resource, 'user'),
-            ],
-        ];
-    }
+    protected $relationships = [
+        'user' => 'uploaded-by',
+    ];
+
 }